Add prod configuration and installation
The configuration for an h2o instance to listen on /var/run/h2oblog.conf
is enabled for production.  The commands included allow me to push an
updated h2o configuration with `make install-prod`. A touch-file is
maintained at .install-prod in the local directory to compare the latest
changes in the config file.