Add dates to blog posts
Using some regexes and the power of sed, we're going to add the date to
blog posts, assuming that you provide an invisible `date` link anchor in
your posts. The 'right way' to do this would be to lean on a library
like pulldown-cmark to parse out those tags. I also don't yet htmlescape
the entities in the title.